+++ src/main/java/com/demo/repository/SprintRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author Mykhailo Balakhon
  * @link <a href="mailto:9mohapx9@gmail.com">9mohapx9@gmail.com</a>
  */
-public interface SprintRepository extends JpaRepository<Sprint, Long> {
+public interface SprintRepository extends JpaRepository<Sprint, Integer> {
 }
